Refactor: clickhandler를 GameDescBox로 올리기
